+++ listings/ch11/listing_09.js
@@ -1,5 +1,26 @@
-export function extractPropsAndEvents(vdom) {
-  const { on: events = {}, ...props } = vdom.props
+export function defineComponent({ render, state, ...methods }) {
+  class Component {
+    #isMounted = false
+    #vdom = null
+    #hostEl = null
+    // --add--
+    #eventHandlers = null // --1--
+    #parentComponent = null // --2--
+    // --add--
 
-  return { props, events }
+    constructor(props = {}/*--add--*/, eventHandlers = {}, parentComponent = null/*--add--*/) {
+      this.props = props
+      this.state = state ? state(props) : {}
+      // --add--
+      this.#eventHandlers = eventHandlers // --3--
+      this.#parentComponent = parentComponent // --4--
+      // --add--
+    }
+
+    // --snip-- //
+  }
+
+  // --snip-- //
+
+  return Component
 }
