+++ listings/ch11/listing_10.js
@@ -1,21 +1,5 @@
-// --add--
-import { extractPropsAndEvents } from './utils/props'
-// --add--
+export function extractPropsAndEvents(vdom) {
+  const { on: events = {}, ...props } = vdom.props
 
-// --snip-- //
-
-function createComponentNode(vdom, parentEl, index, hostComponent) {
-  const Component = vdom.tag
-  // --remove--
-  const props = vdom.props
-  const component = new Component(props)
-  // --remove--
-  // --add--
-  const { props, events } = extractPropsAndEvents(vdom) // --1--
-  const component = new Component(props, events, hostComponent) // --2--
-  // --add--
-
-  component.mount(parentEl, index)
-  vdom.component = component
-  vdom.el = component.firstElement
+  return { props, events }
 }
