+++ listings/ch11/listing_14.js
@@ -2,11 +2,40 @@ export function defineComponent({ render, state, ...methods }) {
   class Component {
     // --snip-- //
 
-    // --add--
-    emit(eventName, payload) {
-      this.#dispatcher.dispatch(eventName, payload)
+    mount(hostEl, index = null) {
+      if (this.#isMounted) {
+        throw new Error('Component is already mounted')
+      }
+
+      this.#vdom = this.render()
+      mountDOM(this.#vdom, hostEl, index, this)
+      // --add--
+      this.#wireEventHandlers() // --1--
+      // --add--
+
+      this.#isMounted = true
+      this.#hostEl = hostEl
     }
-    // --add--
+
+    unmount() {
+      if (!this.#isMounted) {
+        throw new Error('Component is not mounted')
+      }
+
+      destroyDOM(this.#vdom)
+      // --add--
+      this.#subscriptions.forEach((unsubscribe) => unsubscribe()) // --2--
+      // --add--
+
+      this.#vdom = null
+      this.#isMounted = false
+      this.#hostEl = null
+      // --add--
+      this.#subscriptions = [] // --3--
+      // --add--
+    }
+
+    // --snip-- //
   }
 
   // --snip-- //
