+++ listings/ch11/listing_15.js
@@ -1,0 +1,15 @@
+export function defineComponent({ render, state, ...methods }) {
+  class Component {
+    // --snip-- //
+
+    // --add--
+    emit(eventName, payload) {
+      this.#dispatcher.dispatch(eventName, payload)
+    }
+    // --add--
+  }
+
+  // --snip-- //
+
+  return Component
+}
